[societegenerale] handling new errors
Management of unhandled errors in the spge part of the module.
Sometimes in the spge part of the module,
we get the value 'nok' for the key ‘state’ in the json
for different 'reasons', which can be
ENCADREMENT_KYC_PREAVIS, ENCADREMENT_KYC_POST_PREAVIS, INSCRIP_OBL,
in this case we raise ActionNeeded because the user has to fill in
some information on the bank website.
We also manage the ‘err_is’ and ‘err_tech’ cases in which we
raise BrowserUnavailabe as in the browser of the main module.
Sometimes the key 'reason' has the value 'FIABILISATION_TS'.
To manage this case in the main module there is a function named
check_skippable_action_needed.
However, in some cases the function does not seem to be sufficient
to pass FIABILISATION_TS state.
In this case, we have the key 'chgtnivauth' in the returned json.
During login, this key is also present when an SCA is required.
It is likely that, in some cases, an SCA is also required to pass
the 'FIABILISATION_TS' state.
So, in the main module, I add the possibility to trigger
an SCA if the json obtained in the check_skippable_action_needed
function contains the chgtnivauth key.
We do not have a PSU to test the solution regarding FIABILISATION_TS.