[boursorama] Handle identical transactions
For checking accounts, there are some authorizations transactions that
can be exactly the same. Their ID, label, date and amount are identical.
Nothing in the data we parse can help to differentiate them. Since
transaction id is optional and never used in the module, it is now
discarded for coming tansactions. Once they get the status book on the
website and are regular transactions, a unique ID can be found in the HTML
like for any other transactions parsed by iter_history method.