[societegenerale] Add a retry when getting sign info
Getting this page sometimes result in getting a 'nok'
page with an 'err_tech' reason, which is the temporary
error code for Société Générale's website.

We add a retry and ensure that the 'nok' is detected
even if Python is optimized to the point it removes
asserts.