[boursorama] handling new urls
Handling new url for loginpage. In the get_filled_account
function we make requests on urls of this type
/account/sav/cav/<account_id>
and we are redirected to /account/cav/<account_id>/movements
It's not very clear, but sometimes, we are redirected to
/se-deconnect then to /connexion/?deconnexion=
However, this url is not managed as a login url, so at the line
'self.page.fill_account(obj=account)' the code
crashes because page is None.

In other cases the problem comes from the fact that
TNC accounts ("titres non cotés") are not handled for the moment.
This kind of accounts have this type of url
account/tnc/<account_id>/investments
If the user has such an account it causes a crash
for the same reason as before. For the moment I added an empty page
for this kind of accounts that allows us to ignore the account
if the user has some without preventing him to connect his other accounts,
while waiting that we manage this type of account.

Problem reported to the Wealth team.