[cragr] Add retry for accounts_url request
During login process, we get some accounts URL in a JSON
response in order to finalize login. The URL given can sometimes be
unavailable and this is already handled by the module. But this request
can also timeout sometimes. A retry has been added, we'll have to wait
for the issue to happen again to see if it's effective.