[lcl] handling new url
Handling of a new url for the login page. In browser.py in
go_back_from_life_insurance_website we can be disconnected
and then redirected to an unmanaged login url
https://assurance-vie-et-prevoyance.secure.lcl.fr/acces-non-authentifie.
To avoid this I added this url in the login urls.
At the end of the function we check if we are on the login
page and if it’s the case we call do_login.
This part could not be tested because the
disconnection seems relatively random.

There is another problematic case, in the browser.py
in get_accounts function, when we call real_estate.go()
we can be redirected to the "no_perm" page.
This happens for example with users who log in
with a recovery access code (this is a temporary code
that gives restricted access to the site).
he crash occurs because we try to call iter_account
on this page which is not possible.
To avoid this, we add a condition, if we are
on the no_perm page we trigger a logger.warning
instead of trying to call the iter_account
function on this page. This is how this kind of
problem is handled in the rest of the module.