[caissedepargne] Add AppValidation case for check_errors()
check_errors() method is called during the login and checks various
cases of errors. If there are no errors, the JSON we check has its
'phase' key set to one of the authentication method. Until now,
'CLOUDCARD' (which corresponds to an app validation) was not in the list
of values we check. I can only guess but can't confirm that the reason
it is crashing now is because the structure of the JSON response has
changed. For app validation cases, the value we checked was probably in
['response']['status'], which is a case that we already handle.