[lcl] handle authentication error
Sometimes, the redirection to the twofa authentication page
leads to an error page. Retrying the twofa authentication
at least half an hour later solves the problem.

The retry decorator was first introduced to solve that
issue, but it doesn't work. Trying to reach the url again
fails, and it fails with a different error, so the exception
never rose again. Besides, logs show that retrying just after
the initial error doesn't help; one needs to wait a bit longer
for the authentication to work.