[cragr] Add new message for ActionNeeded
New message asking the user to check and/or update his personal
information, can't be skipped by trying to go directly on subsequent
URLs. There's a "skip for later" button but we can't guarantee the
message will be displayed at the next connection for the PSU if we click
on it so raising an ActionNeeded is the best thing to do.