[s2e] handling error message and absence of form on account page
Some accounts are bugged, we cannot access them on the
account page.
In that case, the website returns an error message
that is now handled.
Also, when the account is bugged or when there
are no accounts there is no form on the page which
caused a bug on our side.
Now we first verify if there is a form on the page.
If there is not, we display the corresponding exception