[ldlc] Reworking the module
Urls to navigate are now different from parents.

Par and pro site are now completely different.

Most of the par method were identical to the parent module with
just different URLs so they were removed.
The xpath for documents was updated in case of command with multiple package.

LdlcBrowser was only used by LdlcProBrowser.
The two were merged.