[fortuneo] Handle new wrongpass behavior
Posting credentials is done without allowing redirection, thus allowing
us to control the flow and check if an 2FA is going to be triggered.
That's why wrongpass cases were checked thanks to the location header we
could get after posting the credentials. Now, the response for invalid
credentials cases is just a 401 with no location header attached to it.
To get the related error message, we must manually reload the login
page. That's how the website works, it does not use any location header
for these cases. For 2FA connections, the response is still a 302
with a location header, as it was before.