[freemobile] Handle duplicate id in subscriptions
We use the phone number as id, this is a problem because
several subscriptions can have the same phone number.
On the other hand, each subscriptions has a unique id,
so we replace the current id with this unique id.

I also changed the xpath for retrieving documents because
it was too strict and did not allow retrieving documents on
terminated subscriptions.