[bp] properly implement fill_account
Used to retrieve account.ownership.
This prevents whole iter_accounts to break
when bug is situated on the get_profile method.
Tested on accounts with/out ownership, Par and Pro.
This is a follow up to 3a1a9cb059becfc5affdb9e7ae3d41d47ff47fa7
but make it more resilient.