[boursorama] Handle ULTIM and METAL cards label cases
Boursorama stopped the commercialization of their PREMIER cards and
are replacing it with ULTIM or METAL cards. Unfortunately, their website is
still mixing both names. On the page where we fetch the accounts,
accounts label are - most of the time - something like "CARTE VISA
PREMIER". We must use that label to fetch the card number on another
page. But on that another page, labels are - most of the time - formed
like "CARTE VISA ULTIM". There are some cases where the names are
coherent but it seems to be pretty rare. Label is the only information
we have that will guarantee us to match the right card number, so we
have no choice but to replace "PREMIER" for "ULTIM" or "METAL"  when
the XPath with "PREMIER" is not found.