woob.browser.browsers: save request to HAR file even when an exception occur
Because previous version needed a response object to store entry to HAR
but python-requests doesn't provide any response when an exception
occur, but it could be very useful to have the request that lead to a
timeout or any other exception that contains the request.

Empty response dict is based on what firefox or chromium produce.

Move all code handling HAR in a new dedicated class and file
for better readability and maintainability.