[carrefourbanque] Handle new page which fix login error
Carrefourbanque redirects users onto a kyc verification page,
users have to fill in their name, birthdate and birthplace.
This happens if the user tries to login for the first time
with new login (internet id).

Handle a new page "kyc" which is usefull to
raise an ActionNeeded if we are on this page during login.