[boursorama] Add default value on card_key Coalesce
When a Card Account is closed it is still displayed
on the website for a moment but it no longer has a
card associated. Absence of any value was making
card_key Coalesce crash.
Also, on fallback, when going on CardInformationPage
we end up on the right url but the page only display
an advertissement message about its card subscription