[axabanque] Handle insurances from AxaBanqueBrowser
Axa has three spaces, insurances only, bank only and bank and
insurances. For bank space users, insurances accounts were not fetched
since they were located on the insurance space. Browser switching now
correctly fetches these insurances. The tricky part is that insurance
space can itself have two spaces (metropolitan and overseas
territories). We can call some URL from the AxaBanqueBrowser to
determine on which space we're gonna land. We keep this information in
the state of the browser to pass it to the AxaAssuranceBrowser.