[bp] Make "COMPTE ATTENTE" handling non-case sensitive
"COMPTE ATTENTE" are checking accounts. When fetching their history, they
must be handled differently than other checking accounts. This
distinction is made by looking at their label that was "COMPTE ATTENTE"
until now. "Compte Attente" can now be found but for the moment, it is
impossible to know if the full uppercase label can still be found so the
code handles both cases.