[axabanque] Handle more switching cases between browsers
We can't guarantee the order of iter_accounts. From that order depends
which browsers will be called first (AXABanqueBrowser, AXABourseBrowser
or AXAAssuranceBrowser). Some rare cases were not covered. For example
if the user has a bourse account, then it is possible that we switch to
the bourse browser first. Since no switching method was used in this
bourse browser, if the user had an insurance account, bourse browser
would try to get insurance investments or history through its own
methods.