[societegenerale] Various fixes for wealth accounts
New types of pro loans are fetched since a few days. A few particular
cases were not covered. Some loans must be skipped since they contain
almost no information (Crédit Bail Mobilier or Location Option d'Achat
for example). Pro loan page can sometimes return an unavailable status
even on the website, handling of that message have been added to prevent
the for loop fetching the accounts stopping. Also added a default
NotAvailable for an element of the label that is called
"informationFacultative" and, as its name says, this value of this key
is not always provided in the JSON. Finally, added a condition to skip
some PER and PERCO. These are accounts that have no information on the
website except a label and a contract number. Clicking on them on the
website leads us to an error page.