[amundi] Fix fake and splitted accounts
This fix aims to reproduce how Amundi processes
accounts datas.

eg: a user sees on his dashboard 2 accounts,
but the site returns us 16 accounts.
And all of them are linked to a "real" account,
actually an account that the user sees.

So we link them thanks to the "merge_accounts" function.

We remove the "fakes" ones and ad their balances to reproduce
what the site does, and so what the user sees.

We repeat this behavior for investements too.