[creditmutuel] Fix month retrieving for pagination
An IndexError occurs while retrieving the month needed
to handle the history pagination. This is caused by the fact
that we are wrongly accessing to the second item of the months
list even when its size is equal to one.
Fix the index to always access to first item in the list.