[cragr] Increase timeout to 20.0
A lot of timeout errors are happening since a few weeks. This one is not
tested as I couldn't reproduced it when coding it. But there are for
sure differences in time between two requests, one can be quite slow and
another quite fast. Trying to increase the timeout might reduce these
timeout errors. If this fix has no effect, it will be reverted.