[cragr] Handle tax residency filling page
At the end of the login, some users are asked to fill a tax residency.
There's a button to do it later but we can't do it for the user. Trying
to access accounts will always lead to this page so we have no choice
but to raise an ActionNeeded.