[caissedepargne] Updated raise_otp_emv_authentication method
In some configurations, we could send an SMS while woob is
not interactive. In these cases, the SMS was sent when we
validated the virtual keyboard. To avoid this, we add a
check if an SCA is expected before executing do_vk_authentication
and if yes, we check if woob is in interactive mode.