[amundi] Fix account object has not sub_accounts
Previous commits
do not handle _sub_accounts properly in merge_accounts
function due to a wrong loop organization and naming.

Fix regression:

̀`AttributeError Account has not attribute _sub_accounts`.

This was caused by a for loop that had the same name as
a variable.

To solve that we now fetch the master_account before
merge loops.

It avoids nested loops and allows us to return all
accounts in case there is no master (we are not sure if
it is possible).