[amundi] Fix master account id
As we previously weren't returning the right
accounts and ids and that end users were deactivating
accounts with a balance equal to 0, previous commits
caused an issue regarding matching of activated accounts.

To solve that we simply set the master_account id to the id of the
first positive balance sub account.