[afer] rework bypass cloudflare on login
A previous commit generates 429 quickly if we try to connect
to several accounts at the same time.
To reduce the number of requests,
we do not trigger 403 cloudflare
and go directly to the good page.

To not trigger 403 cloudflare we have to
follow login redirect but by changing
http to https