[orange] Update AssertionError in iter_documents for pro
Since the module exists, we always had a an AssertionError planned for
the day we would reach a total number of 72 documents for any pro
subscriptions. This was done that way because the website always
requests that documents route with a parameter "count" equal to 72. On
the website, if we try to modify this parameter, it returns a JSON with
a number of documents equal to the parameter. Problem is, all
connections won't go further than 5 years of history for the moment
(every connection tested, for the oldest one, starts at 01/2017, that's
72 months from now), so we can't see how the website will react when
there will be 73 months of documents to display. For the moment, fixing
this to get the december months because the actual AssertionError can
only work up to november. This fix will have to be monitored next month
when the website will have connections existing for more than five years
so that we can understand how it deals with these cases.