[freemobile] fix label and recap handling
In the case you have multiple lines, the bills (documents) have
and identical label. In the past, they used to have the phone
number in the label, which made more sense, however a recent
change in web site removed it.
So we add it back.

Additionnaly, there is a summary ("recap") which had an issue -
the id was changing each day, even though the summary was the same.
We now infer the id from the date of the summary, to reduce
the number of ids generated.