[amundi] Fix investment merging
In order to be ISO with the website
we have previously implemented merge account.
Now we have to fix the investment merging on each sub-account.
In some cases,
an investment can be splitted in 2 different accounts and keep the same code.
In order to avoid removing by error a splitted investment,
a `remove duplicate` method has been implemented in browser.py
and is called for each account and not globally.