modules loader: introduces the 'requirements.txt' file in modules
Modules are now distributed commonly to a major version of woob, to be
able to update modules without updating woob.

Repository urls are now:
https://updates.woob.tech/MAJOR/main/

This commit introduces a requirements.txt file in modules directory to
check that the module is compatible with your woob version, and if
dependencies are installed.

For example, if the version of woob is 3.3 and the dependence of a
module is "woob ~= 3.4", it fails to load.

Even when a new version of woob is released, modules can keep a
dependence of woob to a previous version. When it depends on a new
feature of an up-to-date version of woob, the requirements.txt
dependence is updated.