[milleis] Fix iter_investment and iter_history for some accounts
Some wealth accounts have to go to a market space on a different
website. The URL we fetched to go there changed a bit. For
iter_investments, we used the URL "/milleis/trading/positions/security"
which is now "/milleis/trading/positions/realtime". We used to replace
"positions" for "movements" since this was the only part changing to get
the history of the account. Now, we must go to that investment URL and
then request the history URL if we're in iter_history.