[boursorama] Fix label fetching for some investments
Some investments like "Fonds Euro" need a specific XPath. There are some
rare cases where there are two investments of this type that must be
fetched through this XPath. In this case, we had the XPath catching the
names of the two investments, merging two investments names in one
label.