[s2e] Handle new URL for bnp_investments
For bnppere "personeo" connections using s2e browser, we are now
redirected to a new URL for bnp_investments page. This made the browser
think it was not logged if we called for history after investments for
example, since it was unable to handle the page (and it did not crash
because we did nothing with that unhandled page). The browser trying to
relogin while it was already logged triggered a strange bug on the
website where it made us land on a page asking the user to accept his
terms of conditions (but nowhere on the user space this page really
exists, plus if we get on this page, we can just go on any other page
without problem and the page never appears again). Added the new URL
in new_bnp_investments since the old one is still needed
for investments details matching (the "home" of bnp_investments is
now "personeo.epargne..." but to get the details we have to go through
some "optimiser.epargne..." URLs like before). Also added some details
about a 401 that was already here when requesting investment details,
this happen randomly on any given investment, couldn't find a way to
prevent this.