[lcl] Optimize iter_investment of BoursePage
The iter_investment method doesn't work well with
child module. In fact valuation and unit_price
were inverted.

That was due to the parsing made with ListElement
which was terrible. Using TableElement is cleaner
and fixes the compatibility with the child.

I also modify some conditions that had no sense.