[apivie] Updated signature_key managment
When we submit the creds, we need to create a jwt with
a signature key that we can reconstruct from the js on
the page. The address of this js is not consistent
over time but the key seems to stay the same. So it is
more efficient to hardcode the key. I added a logger
to warn us if the key is not valid anymore.