[milleis] Update URLs and add fill_loan
Most of the URLs have changed and some of them have a new param for
account id instead of the accound id in the path. Loans were fully
fetched in one JSON, now the main informations are still in the old JSON
route but the details are located on another route, hence the fill_loan.