[themisbanque] Updated get_operations to avoid duplicate transactions
Sometimes the website returns the same list of transactions
for each history page. So we updated the next_page method
of get_operations to prevent duplicated transactions.

It replaces _dedup_transactions method. This old method
stopped the iteration if a transaction had a more recent
date than the date of the last transaction. But it doesn't
work if all the transactions on the page that the bank
sends in a loop have the same date.