[axabanque] Handle 401 errors
If the user
tried to synchronize within 5 minutes after a first synchronization, he
would get a 401 because authorization token would not be set in the
header (it is only set after regular login or SMS handling and not kept
in the state). After 5 minutes, the STATE_DURATION would make the state
expire and the user reconnect, making him have a new fresh token.
Updated the STATE_DURATION to 10 minutes like it really lasts on the
website and also skipped locate_browser to avoid loading URLs that must
be accessed with the token, provoking 401 errors. Authorization could be
stored in the state but then it would only be useful for users that need
to do two synchronizations in less than 10 minutes. Instead, if they
have to do this, module will relogin. refresh_access_token from parent
allianzbanque can't be used for axabanque.