[axabanque] Handle random failed redirections
AxaBanqueBrowser must go to a redirection bouncer to know if the user
has his insurances on regular space or on outremer space. This bouncer
is supposed to redirect us to specific URLs that allows us to know which
space is used. Sometimes, it randomly gets the user to
AxaAssuranceBrowser homepage instead of AxaAssuranceBrowser insurances
pages. It can even happen when browsing on a regular browser. Going to
insurances a second time seems to be enough to get the real redirection.