[caissedepargne] Rework module
The whole point of this rework is to drop the old caissedepargne space
that is still online but is missing more and more features each day. We
already supported the new login but this has been extended for all
connections and reworked. A new authorization process at the end of the
login is necessary to get the right token that allow browsing the new
space. The parsing is way simpler since we get all the data from
caissedepargne "new" internal API. Unfortunately, a session still can't
last for more than around ten minutes and the 2FA handling is
problematic at the moment since the device won't be remembered, making
the 2FA asked at each login.