[s2e] ignoring useless URL
The URL:
'https://www.comgest.com//fr/fr/professionnel/...&tab=12'
is automatically opened when we try to retrieve
the investments of certain accounts.
However, this URL leads us to an error page
with a 500 status code.
Since there is no useful information on the
Comgest website, we ignore this URL.