[banquepopulaire] Update user_types handling
When posting the username on login_api, we get a JSON response with
enough data to determine the type of that user (part, pro, etc.). That
data was given as a label like "abonnement particulier" for "part"
users. Now, the label is always empty and the only way to match the user
type is to use the "code" value instead of the "label" one. There are
still some unexplicit labels in another part of the JSON but they are
unusable since "PM" (presumably "Personne Morale") is used for "ent"
and "pro".