[boursedirect] Implement 2FA
All connections that have activated 2FA (SMS or TOTP) have now an error
500 at the login. Some header must be specified to avoid this. Specifity
of the 2FA is that a user can have both factors activated. The website
choses to present the TOTP if that's the case so the code is following
this behavior. Plus, TOTP means less requests (we do no have to make the
request to generate and send the SMS OTP) and if something goes wrong
because of a change on the website, TOTP does not trigger any
notification for the user, unlike the SMS. device_id cookie must be used
for subsequent logins once the 2FA has been done a first time and it
seems to be valid for approximately 13 months.