[boursedirect] Improve device_id cookie creation
If we already did the 2FA but have to do the login again, and if, at the
same time, the cookies can't be loaded or are expired, then we recreate
the device_id cookie from the device_id value kept in the state to avoid
triggering the 2FA again. The way it was done before could make
duplicates cookie. Added a condition and created the cookie in a cleaner
way. Also improved the comment above that wasn't explicit enough.