[boursorama] Fix amount attribute for SavingMarketPage's iter_history
The table for SavingMarketPages's iter_history is a bit broken, it has
some invisible td elements that makes TableElement unable to match some
columns name with the right row. For example, to match "Montant" column,
we would have to use the "Quantité" column instead. For the moment,
Dropped the use of TableCell for the amount attribute.
Boursorama made the change, then reverted it, then switched back to it.
Added a Coalesce to handle both cases.