[creditmutuel] handle deferred loans
In the case of a deferred loan, we now return the monthly
franchise payments in next_payment_date
and next_payment_amount.

If the loan is in the franchise status,
then it is deferred.

The repayment_start_date can only be determined
if repayment has not yet begun.

In this case, we can find the information
"xxx sur xxx mois" on the bank's website,
which is equal to "nb_payment_left over duration".
If they are equal, then next_payment_date is
equal to repayment_start_date.