[banquepopulaire] Retrieve loan's fields for all loan accounts
Until now, we didn't retrieve loan fields for some loan accounts.
This was due to a sneaky case:
A specific title of an html tag ('Le détail de ce contrat')
was missing in the html for some loan accounts.
And previously, we only get loan fields if this title
of html tag was present.

After few tests, it appears that we can get loans' details
even when this html tag is missing.

Now, we only test if the account is a loan/mortgage,
and we throw the _has_details attribute which is useless now.

Moreover, in browser.py, some loan accounts didn't pass
through the loop where get_loan_from_account method was
executed.

Instead, these specific loan accounts are passing through
another loop in which I added a call to get_loan_from_account method.

It is not explicit in the code, but get_loan_from_account
method will not be executed twice for the same loan account.