[filters/standard] Handle scientific notation with CleanDecimal
When input is in scientific notation,
CleanDecimal raised NumberFormatError with
'There should be exactly one number to parse' because
self.matching matchs left and right parts of 'E' char.

Now, thanks to is_notation_scientific regex,
we can detect if the input is a scientific notation
and can continue the process.