[bp] Handle more cases of 2FA messages
Recently, website changed a bit the page where the messages about 2FA
activation are displayed. They added a lot of different cases for the
div name that displays what we need to check. It always start with
"DSP2_xxx" and only the div that has the "mtm" id is of interest.