[caissedepargne] Handle SiteSwitch to old CE for PP connections
Even though PP connections' accounts can be found in a JSON on new CE
space, they should not be scrapped from this new space. No details
whatsoever can be found there, meaning no history or investments. This
is because CE never migrated these PP profiles to the new CE space. It
is impossible for these users to go to the new interface, unlike any
other type of profile. Login for PP is very similar to any other login,
except we don't have to do the last authorization process and we can
directly be redirected on the old space. From there, SiteSwitch switches
the browser for OldCaisseEpargneBrowser.