[amundi] Add token_header to the states
Recently, parent module amundi was updated to handle 2FA. Locate browser
is silently crashing with a 401 most of the time, making the module
login again from scratch. This is because token_header is not in the
state while some URLs need it or the response will be a 401. Adding it
to the state to keep the value between synchronizations. caels child
module is crashing in a non-silent way because the 401 is happening in
iter_accounts (locate_browser loads another URL and everything is fine,
so the 401 is not happening in it and is not catched). Anyway, wether
this is in locate_browser for amundi or iter_accounts in caels, the root
of the issue is the same, we have to provide token_header when needed.