[caissedepargne][cenet] Fix deferred card history fetching
Module has been updated a few months ago to handle cards that are not
displayed on the website but have deferred transactions listed on the
website. There are also regular cards that are displayed on the website
and had, until now, their transactions linked to the history of the
parent account. We were not fetching transactions of these deferred card
since their history is now on the same route than for cards not
displayed on the website (apparently, could not test any of these cases
that seem to be rare). Anyway, changed the condition so that we get
proper deferred card history for any card scenario and distinguish it
from regular checking account (the method fetching the history can take
a regular checking or a checking used for a card to get a few data
first).