[cmes] fix remote disconnected
In cmes and its child components,
we encounter a random bug on the investment
 details route,causing requests to succeed
 intermittently.
To prevent connection failures,
we implemented a mechanism to detect and handle
the error when it occurs, allowing us to continue
without crashing the module.
This ensures that we can still obtain other information.
Additionally, a logger has been implemented
to collect as much data as possible
for comparison and to explore potential
correlations in an effort to find a definitive solution.

By taking these measures,
we not only prevent the module from crashing
 on the details route but also increase
 the likelihood of obtaining these details during
the next synchronization, as the bug occurs randomly,
with requests sometimes succeeding and other
times failing for unknown reasons.