[caissedepargne] Fix obj_raw behavior
Some transactions have a date in their raw labels that we use for
FrenchTransaction patterns. If there is a date in the raw attribute, at
some point FrenchTransaction will use parse_with_patterns method. If the
year is missing in the raw attribute, this method will try to use the
date attribute of the object. The date attribute will be NotLoaded if
obj_date has been redefined, like it has been done a few days ago for
caissedepargne, because redefined attributes values are set after
regular attributes values. This made every transaction that has a date
in its raw attribute crash. Redefining obj_raw so that it's executed
with redefined obj_date.