[delubac] Fix CleanDecimal for accounts' balances
Recently, there were a few changes on the website, especially around
iter_accounts part. The last fix was tested on a few connections that
all had accounts with balances inferior to 1K. In those cases, balance
is displayed like "40.88". Last fix thus changed CleanDecimal filter
from US to SI but in fact, US is still the notation used on the website
(for an amount above 1K, this looks like "1,058.37"). Putting back US
filter for balances.