[boursorama] Correctly type the accounts labeled "pea_pme"
The fact that we only check if the key contained
in ACCOUNT_TYPES is present in the label
can lead us astray, as is the case here where
an account is labeled "PEA_PME PELLEREAU."

When we perform the MapIn, Woob stops at the key "pel"
(type saving) because it recognizes that "pel" is contained
in "PELLEREAU."

However, it should have been typed as a "pea" type.
To temporarily fix this, we arranged the keys so that "pea"
comes before "pel" allowing "pea" to match before "pel"
and thus the account will be correctly typed.