[ticketscesu] add CapDocuments, drop CapBank, module upgrade
Module not tested with credentials but basic existing
login and CapBank methods can seemingly be used.
They are updated to the last Woob standards.
But CapBank is dropped nonetheless as it is not falling under
any use case at the moment.

CapDocument is added in preparation of its methods to be implemented.
Consequently this module serves as an empty shell to be filled up
once connections are made on it.