[s2e] Handle bnp_investments timeout
URL to get bnp's investments details has been crashing
randomly for years with random 503. This is still the
case except that now, on the website, this 503 takes
up to two minutes to be triggered. The module's timeout
set at 60 isn't even enough for this. Rather than making
very long synchronization, we now assume that 30 seconds
timeout on this specific request is enough to consider
the details as unavailable.