[freemobile] Rework authentication flow with added 2FA
Authentication flow has changed, and now requires a 2FA code - mostly sent to your
account SMS number - but can also be sent by mail (when your account has been ported
to another provider for example).

Note that a new boolean parameter `force_twofa_type_email` is added (default: `false`).
When set to `true`, it forces a new 2FA to be sent via email - just after the one that
was sent by SMS (which becomes invalid).
(Unfortunately I could not find a way to completely remove the first SMS 2FA)

The rest of the module is untouched.