[societegenerale] fix MFA retry not properly handled
It seems that a recent change in worklow / non-redirection of page impacts
our ability to follow a proper MFA workflow.
This triggers the following error:
```
Bug(societegenerale): 'donnees'
```
By resetting the current page when asking the MFA, we force woob to detect
then unauthenticated state and thus handle the MFA that was entered by the user.