fix: Simplify woob.exceptions.__getattr__
This function is meant to customize Python object discovery
when importing the module. It is used as a fallback when an object is
not found through the regular module.__getattribute__ lookup [1].

Since the function does not return an exception but raises it,
make the code exit early when the attribute is not in the list of
deprecated classes and make it clearer that it returns an object
by having the return outside of any nested blocks [2].

Simplify the split to spare some CPU cycles and compact the exception
handling since it generates the same outcome.

[1] https://docs.python.org/3/reference/datamodel.html#customizing-module-attribute-access
[2] https://docs.python.org/3/library/typing.html#typing.NoReturn