refactor: Transaction filtering in oney module using CapBank.iter_transaction signature
While there is no requirement to use it as it does not apply to
the same class, it is probably better to keep a sense of how CapBank
is trying to achieve transaction listing. The Transaction objects
creation condition are refactored to account for the new parameters and
the previously set coming attribute.