fix: Improve recipient matching for transfer operations
Rework matching with clear cases and add lookup in transfer history
which contains high quality data.

Rename property exposing the lookup for use by OFX export to be more
inline with project's naming. It was not in use yet so no issue. Maybe
it would be worth using already public `Transaction.counterparty`.