test: Add Revolut transaction parsing example
Instant transfer have a different form than regular transfer. The line
contains the name of the beneficiary instead of the personal alias and
the BIC is in BIC11 form, not BIC8. Unfortunately, this exposes some
limitations in schwify.

The example fails on purpose to illustrate the issue until a fix or a
workaround is implemented.

Refs: https://github.com/mdomke/schwifty/issues/242